+++ database/migrations/2023_04_08_215856_create_expenses_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('description');
             $table->string('purpose');
             $table->date('date');
-            $table->integer('value');
+            $table->double('value');
             $table->timestamps();
         });
     }
