+++ database/migrations/2023_04_08_220156_create_invoices_table.php
@@ -21,12 +21,11 @@ return new class extends Migration
             $table->date('date');
             $table->integer('remaining_amount');
             $table->foreignId('order_id')->references('id')->on('orders');
-            $table->integer('value');
+            $table->double('value');
             $table->double('discount');
             $table->double('tax');
             $table->double('total');
             $table->string('massage');
-
             $table->timestamps();
         });
     }
